# Feat: Update indexing of parquet dataset and also add streaming support to huggingface datasets