## What does this PR do?

This PR introduces the following improvements:

- [x] Updates `HFDownloader` to use `hf_hub_download`, which can be combined with `hf_transfer` for faster downloads.  
- [x] Enhances the parquet dataset indexing process to retrieve metadata and build index without downloading entire files.  
- [x] Adds a low-memory usage option to the Parquet item loader.  
   -  Note: When `low_memory=True`, shuffling has a significant performance impact. It is not recommended to use shuffling with low-memory mode until the next update.
- [x] Updates test cases to reflect these changes.  

Fixes  #502.

### Example usage:
```python
from tqdm import tqdm
from litdata import StreamingDataLoader, StreamingDataset
from litdata.streaming.item_loader import ParquetLoader

# Create a StreamingDataset object
dataset = StreamingDataset(
    input_dir="hf://datasets/open-thoughts/OpenThoughts-114k/data",
    item_loader=ParquetLoader(low_memory=True), # by default it has low_memory=True
)

data_loader = StreamingDataLoader(dataset=dataset, batch_size=8, num_workers=4, shuffle=False, drop_last=True)

for i, batch in enumerate(tqdm(data_loader, desc="Streaming data")):
    pass
```

## Benchmarks  
### Using LitData
These benchmarks were generated using this [script](https://github.com/bhimrazy/litdata-benchmark/blob/main/stream_hf_dataset.py) with the following settings: `batch_size = 256`, `num_workers = 32`, and `machine = A10G`. The results may vary slightly across different runs.  

> For more detailed logs (older), please check [this comment](https://github.com/Lightning-AI/litdata/pull/505#issuecomment-2734439804). 

> **Notes:** 
> 1. Currently, enabling shuffle in low-memory mode degrades performance. However, using `low_memory=True` and `shuffle=False` allows streaming of much larger datasets at better speed.  
> 2. Time in brackets represents the time taken to complete that epoch.  

---

#### Dataset: [OpenThoughts-114k](https://huggingface.co/datasets/open-thoughts/OpenThoughts-114k) (3.55 GB)  

| CASE | Low Memory | Pre-load Chunk | Shuffle | Samples/sec 1st epoch | Samples/sec 2nd epoch | Peak Memory Usage (GB) |
|------------|------------|---------------|------------|------------------------|------------------------|----------------------|
| a    | yes        | no           | no     | **14,329** *(7.95s)*   | **15,948** *(7.14s)*   | ~13 GB               |
| b    | yes      | no           | yes     | **14,093** *(8.08s)*   | **16,026** *(7.11s)*   | ~13 GB               |
| c    | no         | no         | no       | **12,145** *(9.38s)*   | **12,779** *(8.91s)*   | ~23 GB               |
| d    | no          | no         | yes      | **11,696** *(9.74s)*   | **12,747** *(8.93s)*  | ~23 GB               |
| e    | no          | yes         | no      | **11,707** *(9.73s)*   | **10,602** *(10.75s)*  | ~33 GB               |
| f    | no        | yes          | yes     | **11,296** *(10.08s)*  | **10,554** *(10.79s)*  | ~34 GB               |

---

#### Dataset: [fineweb-edu (10BT Sample)](https://huggingface.co/datasets/HuggingFaceFW/fineweb-edu/tree/main/sample/10BT) (~26 GB)  

| CASE | Low Memory | Pre-load Chunk | Shuffle | Samples/sec 1st epoch | Samples/sec 2nd epoch | Peak Memory Usage (GB) |
|------------|------------|---------------|------------|------------------------|------------------------|----------------------|
| a    | yes         |  no          | no      | **36,918** *(261.98s)* | **44,803** *(215.87s)* | ~18 GB               |
| b    | yes         |  no          | yes      | **34,983** *(276.47s)* | **39,966** *(242s)* | ~58 GB               |
| c    | no          | no             | no       | - (crashed)            |                        | ~120 GB              |

------------------------
### Using huggingface datasets streaming
These benchmarks were generated using this [script](https://github.com/bhimrazy/litdata-benchmark/blob/main/stream_hf_iterable_dataset.py) with the following settings: `batch_size = 256`, `num_workers = 32`, and `machine = A10G`. The results may vary slightly across different runs.  
> For `num_workers`, it doesn't not seem to accept a value greater than num_shards.
>Warning: Too many dataloader workers: 32 (max is dataset.num_shards=6). Stopping 26 dataloader workers.

##### Dataset: [OpenThoughts-114k](https://huggingface.co/datasets/open-thoughts/OpenThoughts-114k) (3.55 GB)  
| Shuffle | Samples/sec 1st epoch | Samples/sec 2nd epoch | Peak Memory Usage (GB) |
|---------------|------------------------|------------------------|----------------------|
| no | **14,834** *(7.66s)*   | **15,115** *(7.51s)*   | ~9 GB            |
| yes | **12,173** *(9.33s)*   | **12,288** *(9.23s)*   | ~9 GB            |

#### Dataset: [fineweb-edu (10BT Sample)](https://huggingface.co/datasets/HuggingFaceFW/fineweb-edu/tree/main/sample/10BT) (~26 GB)  
| Shuffle | Samples/sec 1st epoch | Samples/sec 2nd epoch | Peak Memory Usage (GB) |
|---------------|------------------------|------------------------|----------------------|
| no | **44,519** *(217.23s)*  | **44,178** *(218.90s)*   | ~55 GB            |
| yes | **33,147** *(291.74s)*  |  **33,900** *(285.27s)*   | ~55 GB            |


## PR Review  

Community members are welcome to review this PR once tests have passed.  
If your PR was not previously discussed in GitHub issues, there's a high chance it will not be merged.  

## Did you have fun?  
Yes, I did.  😊