@tchaton 2025-03-11
@bhimrazy I wonder if we could benchmark pyarrow vs polars for streaming the data