@bhimrazy 2025-03-11
> @bhimrazy I wonder if we could benchmark pyarrow vs polars for streaming the data

Sure will check that.
duckdb also seems to be another option but none of them seem to provide efficient way of reading particular rows.